Update contact form
The contact form is using formfarm currently, which appears to be failing. Their website is basically non-functional, and half of it is now in Russian. Chrome also warned me that the website wasn't safe when I tried to sign in. Using formspree now which I use on my personal websites. Should work much better.